requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
idna==3.2
Pillow==8.3.2
pyTelegramBotAPI==4.0.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
